scripts/requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pillow==10.2.0
preshed==3.0.9
prettytable==3.9.0
pydantic
pydantic_core
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece
smart-open==6.4.0
soupsieve==2.5
spacy
spacy-legacy
spacy-loggers
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.0
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
PyGithub
openai
python-slugify